01_scrapping/requirements.txt
prefect==2.7.7
pydantic==1.10.11
prefect-aws[S3]
python-dotenv
pandas
# pypdf2
# llama-index
# llmsherpa
# llama-hub
deepsearch-toolkit

# prefect-gcp[cloud_storage]==0.2.3
# pandas==1.5.2
# prefect-sqlalchemy==0.2.2
# prefect-github
# protobuf==4.21.11
# pyarrow==10.0.1
# pandas-gbq==0.18.1
# psycopg2-binary==2.9.5
# sqlalchemy==1.4.46
# Unidecode
